View/edit preferred terminologies
